functions/generate_password/requirements.txt
psycopg2-binary==2.9.7
cryptography==41.0.7 